+++ src/main/scala/beam/sim/population/PopulationScaling.scala
@@ -37,7 +37,7 @@ class PopulationScaling extends LazyLogging {
     // check the additional population required to be generated (excluding the existing population)
     val additionalPopulationRequired: Long = totalPopulationRequired - scenario.getPopulation.getPersons.size()
     // check the repetitions(of existing population) required to clone the additional population
-    val repetitions = math.floor(additionalPopulationRequired / scenario.getPopulation.getPersons.size()).toInt
+    val repetitions = math.floor(additionalPopulationRequired.toDouble / scenario.getPopulation.getPersons.size().toDouble).toInt
     // A counter that tracks the number of population in the current scenario (stop as soon as this counter hits the required total population)
     var populationCounter = scenario.getPopulation.getPersons.size()
     val existingHouseHolds = scenario.getHouseholds.getHouseholds.asScala.toSeq
